Fix README to match code
While refactoring the old code to form this module,
the 'uncompressed' attribute and method should have been
transformed to 'compress' and 'toCompressed' string.
While this change has been done in the README, I forgot
to do it in the code.

The change is breaking change, therefore it will be done
only in the next major version 2.0, while in 1.x just the
README is adjusted.